+++ LegendWeathers/WeatherSkyEffects/BloodSkyEffect.cs
@@ -11,6 +11,9 @@ namespace LegendWeathers.WeatherSkyEffects
         private readonly List<GameObject> spawnedEffectsObjects = new List<GameObject>();
         private readonly List<Vector3> effectObjectsPositions = new List<Vector3>();
 
+        private readonly bool eclipsedMusicEnabled = false;
+        private GameObject? eclipsedObjectMusic;
+
         public readonly static float bloodMoonSunAnimatorFixedTime = 0.15f;
 
         public BloodSkyEffect() : base(Plugin.instance.bloodSunObject)
@@ -19,6 +22,7 @@ namespace LegendWeathers.WeatherSkyEffects
             sunlightColor = new Color(1, 0.36f, 0.7f);
             sunSizeFactor = Plugin.config.bloodMoonSizeFactor.Value;
             maxSpawnedEffectObjects = Plugin.config.bloodMoonEffectsAbundance.Value;
+            eclipsedMusicEnabled = Plugin.config.bloodMoonAmbienceMusicType.Value == "Eclipsed";
         }
 
         public override void OnEnable()
@@ -26,6 +30,14 @@ namespace LegendWeathers.WeatherSkyEffects
             if (WeatherRegistry.WeatherManager.IsSetupFinished)
             {
                 base.OnEnable();
+                if (eclipsedMusicEnabled)
+                {
+                    ActivateEclipsedMusic(true);
+                }
+                if (IsEffectActive && (Plugin.config.bloodMoonAmbienceMusicVolume.Value != 1f || eclipsedMusicEnabled))
+                {
+                    CustomizeAmbientMusicVolume();
+                }
                 if (effectObjectsPositions.Count == 0)
                 {
                     for (int i = 0; i < maxSpawnedEffectObjects; i++)
@@ -59,6 +71,10 @@ namespace LegendWeathers.WeatherSkyEffects
             if (WeatherRegistry.WeatherManager.IsSetupFinished)
             {
                 base.OnDisable();
+                if (eclipsedMusicEnabled)
+                {
+                    ActivateEclipsedMusic(false);
+                }
                 foreach (var terrainObject in spawnedEffectsObjects)
                 {
                     if (terrainObject != null)
@@ -71,6 +87,7 @@ namespace LegendWeathers.WeatherSkyEffects
         public void ResetState()
         {
             effectObjectsPositions.Clear();
+            eclipsedObjectMusic = null;
         }
 
         public static void CheckAndReplaceMaterial()
@@ -83,6 +100,27 @@ namespace LegendWeathers.WeatherSkyEffects
                 return;
             }
             bloodSun.GetComponent<MeshRenderer>().material = bloodManager.GetComponent<BloodMoonManager>().bloodMoonTextureMaterials[1];
+        }
+
+        private void ActivateEclipsedMusic(bool activate)
+        {
+            if (eclipsedObjectMusic == null)
+            {
+                eclipsedObjectMusic = FindObjectOfType<EclipseWeather>(true).gameObject;
+            }
+            eclipsedObjectMusic.SetActive(activate);
+        }
+
+        private void CustomizeAmbientMusicVolume()
+        {
+            if (spawnedSun != null)
+            {
+                foreach (var audio in spawnedSun.GetComponentsInChildren<AudioSource>())
+                {
+                    if (audio != null)
+                        audio.volume = eclipsedMusicEnabled ? 0f : Plugin.config.bloodMoonAmbienceMusicVolume.Value;
+                }
+            }
         }
     }
 }